# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Final Project: Tiketku",
    "description": "Platform ini merupakan tempat beli dan booking tiket penerbangan secara online, baik one way (sekali penerbangan) dan round trip (pulang pergi). Platform ini membuka dan menyediakan berbagai jenis penerbangan domestik dan mancanegara. Flow / alur platform dibebaskan berdasarkan kreasi tim final project. Terdapat 1 (satu) peran User yaitu login sebagai buyer. Buyer dapat melakukan pencarian, pemilihan, dan melihat histori pemesanan dan detail tiket.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "brahmastabagus@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://be-tiketku-production.up.railway.app/api/v1",
      "description": "Production server"
    },
    {
      "url": "https://743b-103-162-237-17.ngrok-free.app/api/v1/",
      "description": "Stagging server"
    },
    {
      "url": "http://localhost:8080/api/v1/",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/information/{id}": {
      "get": {
        "tags": [
          "Information Flight"
        ],
        "summary": "Get information flight",
        "description": "Get information flight by id.",
        "operationId": "getInformationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of information to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get informations by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Informations"
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoData"
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Information Flight"
        ],
        "summary": "Update information flight",
        "description": "Update information flight data.",
        "operationId": "updateInformation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of information to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated information object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Information"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Information Flight"
        ],
        "summary": "Delete information flight",
        "description": "Delete information flight by id.",
        "operationId": "deleteInformationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of information to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
    "/notification/{id}": {
      "get": {
        "tags": [
          "Notification"
        ],
        "summary": "Get notification",
        "description": "Get notification by id.",
        "operationId": "getNotificationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of notification to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get notifications by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notifications"
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoData"
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Notification"
        ],
        "summary": "Update notification",
        "description": "Update notification data.",
        "operationId": "updateNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of notification to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated notification object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Notification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update notification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Notification"
        ],
        "summary": "Delete notification",
        "description": "Delete notification by id.",
        "operationId": "deleteNotificationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of notification to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete notification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Invalid input supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res404"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Admin11#"
          },
          "first_name": {
            "type": "string",
            "example": "saya",
            "description": "Nama Lengkap"
          },
          "last_name": {
            "type": "string",
            "example": "user",
            "description": "Nama Keluarga"
          },
          "phone_number": {
            "type": "integer",
            "example": "08123456789"
          },
          "google_id": {
            "type": "integer",
            "example": "1"
          },
          "role": {
            "type": "string",
            "example": "admin",
            "enum": [
              "admin",
              "user"
            ]
          },
          "otp": {
            "type": "integer",
            "example": "12345"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          }
        }
      },
...
      "Information": {
        "type": "object",
        "properties": {
          "flight_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Bagasi 2kg"
          }
        },
        "required": [
          "flight_id",
          "name"
        ]
      },
      "Informations": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "flight_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Bagasi 2kg"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          }
        }
      },
      "Notification": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "example": "Notification",
            "enum": [
              "Notification",
              "Promosi"
            ]
          },
          "title": {
            "type": "string",
            "example": "Transaksimu belum dibayar"
          },
          "desc": {
            "type": "string",
            "example": "Segera lakukan pembayaran pada transaksi anda."
          },
          "user_id": {
            "type": "integer",
            "example": 1
          }
        },
        "required": [
          "tag",
          "title",
          "desc",
          "user_id"
        ]
      },
      "Notifications": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "tag": {
            "type": "string",
            "example": "Notification",
            "enum": [
              "Notification",
              "Promosi"
            ]
          },
          "title": {
            "type": "string",
            "example": "Transaksimu belum dibayar"
          },
          "desc": {
            "type": "string",
            "example": "Segera lakukan pembayaran pada transaksi anda."
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "user": {
            "$ref": "#/components/schemas/Users"
          }
        }
      },
      "NoData": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "data": {
            "type": "array",
            "example": [ ]
          }
        }
      },
...
      "Res404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "URL tidak ditemukan"
          }
        }
      },
      "Res400": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "\"name\" is required"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}